feat: multi-year P95 composites, monthly QA climatology, and destriping groundwork
Multi-year windows:
- ProcessingJob gains optional end_year + window_label; datetime_range spans
  the window and P95 is pooled across all its scenes (models.py, pipeline.py).
- Storage/commits keyed on window_label (storage.py, job.py).

QA count -> 12-month climatology:
- qa_count is now (month, y, x) uint8 (month M = valid obs pooled across the
  window), replacing the single annual int16->uint16 count (pipeline.py,
  zarr_writer.py). Coarsened per-month; exported as a 12-band monthly COG.

Improved masking + de-striping groundwork:
- create_qa_mask adds dilated-cloud (bit 1) + cirrus (bit 2); convert_to_celsius
  clamps to settings.lst_valid_min/max (default -50/80C) to drop ~-124C DN=0
  resampling artifacts (qa.py, config.py).
- New reusable cog.py: single-band LST COG + 12-band monthly QA COG.
- config: load_chunk_size (cuts P95 quantile memory); frisky optional extra.

Scripts:
- pergamino_multiyear_decision.py (1/3/5yr driver: Zarr + COGs + metrics report,
  Frisky hijack behind a Dask fallback / --no-frisky).
- season_aware_p95_test.py (chosen de-striping prototype + synthetic selftest),
  plus percentile / masked / normalized diagnostics.

Docs:
- docs/findings-destriping-and-multiyear.md (full investigation + rejected
  approaches), docs/adr/005-multiyear-monthly-qa-and-destriping.md, README +
  CLAUDE updates. Season-aware normalization is documented as a PROTOTYPE, not
  yet wired into the production pipeline.

Tests: updated for 3D uint8 qa_count, window_label, dilated/cirrus bits, and the
plausibility clamp; new tests/integration/test_cog.py. 118 unit + offline
integration passing.

+++ tests/integration/test_full_tile.py
@@ -228,9 +228,9 @@ class TestFullTileIntegration:
         assert "lst_p95" in ds_read.data_vars
         assert "qa_count" in ds_read.data_vars
 
-        # Check dtype (should be uint16 after encoding)
+        # Check dtype after encoding: LST uint16 DN, QA count uint8.
         assert ds_read["lst_p95"].dtype == np.uint16
-        assert ds_read["qa_count"].dtype == np.uint16
+        assert ds_read["qa_count"].dtype == np.uint8
 
         # Check encoding attributes preserved
         assert ds_read["lst_p95"].attrs["lst_scale_factor"] == 0.01
